refactor: change folder structure to accommodate compressor repo
feat: wip
