+++ topo_processor/factory/__init__.py
@@ -1,0 +1,4 @@
+from .collection import Collection
+from .collection_factory import create_collection
+from .data_type import DataType
+from .item import Item
