+++ topo_processor/metadata/loaders/metadata_loader.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from topo_processor.metadata.item import Item
+from topo_processor.factory.item import Item
 
 
 class MetadataLoader(ABC):
