+++ topo_processor/metadata/loaders/metadata_loader_checksum.py
@@ -1,4 +1,4 @@
-from topo_processor.metadata.item import Item
+from topo_processor.factory.item import Item
 from topo_processor.util.checksum import multihash_as_hex
 from topo_processor.util.tiff import is_tiff
