+++ topo_processor/metadata/loaders/metadata_loader_imagery_historic.py
@@ -2,8 +2,8 @@ import csv
 import os
 from typing import Dict
 
-from topo_processor.metadata.data_type import DataType
-from topo_processor.metadata.item import Item
+from topo_processor.factory.data_type import DataType
+from topo_processor.factory.item import Item
 from topo_processor.util import is_tiff
 
 from .metadata_loader import MetadataLoader
