+++ topo_processor/metadata/loaders/metadata_loader_tiff_test.py
@@ -1,9 +1,9 @@
 import asyncio
 import os
 
-from topo_processor.metadata.collection import Collection
-from topo_processor.metadata.data_type import DataType
-from topo_processor.metadata.item import Item
+from topo_processor.factory.collection import Collection
+from topo_processor.factory.data_type import DataType
+from topo_processor.factory.item import Item
 
 from .metadata_loader_tiff import MetadataLoaderTiff
